// src/clients/generated/commonApi/models/updatePaymentRequest.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * order-manager
 * API order-manager
 * OpenAPI spec version: 1.0.0
 */
import type { PaymentFieldsPatch } from './paymentFieldsPatch';

export interface UpdatePaymentRequest {
  patch: PaymentFieldsPatch;
  /**
   * Идентификатор платежа.
   * @maxLength 36
   */
  paymentId: string;
}